+++ src/main/kotlin/net/perfectdreams/loritta/helper/utils/slash/DailyCheckExecutor.kt
@@ -40,7 +40,7 @@ class DailyCheckExecutor(helper: LorittaHelper) : HelperSlashExecutor(helper) {
 
         val dailies = transaction(helper.databases.lorittaDatabase) {
             Dailies.select {
-                Dailies.receivedById inList users.map { it.id.value }
+                Dailies.receivedById inList users.map { it.id.value.toLong() }
             }.orderBy(Dailies.id, SortOrder.DESC)
                 .toList()
         }
